requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements.txt pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   starlette
    #   watchfiles
beautifulsoup4==4.14.2
    # via yfinance
black==25.9.0
    # via ai_agents (pyproject.toml)
certifi==2025.10.5
    # via
    #   curl-cffi
    #   requests
cffi==2.0.0
    # via curl-cffi
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   black
    #   uvicorn
curl-cffi==0.13.0
    # via yfinance
fastapi==0.119.0
    # via ai_agents (pyproject.toml)
frozendict==2.4.6
    # via yfinance
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
iniconfig==2.1.0
    # via pytest
multitasking==0.0.12
    # via yfinance
mypy==1.18.2
    # via ai_agents (pyproject.toml)
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
numpy==2.3.4
    # via
    #   pandas
    #   yfinance
packaging==25.0
    # via
    #   black
    #   pytest
pandas==2.3.3
    # via yfinance
pathspec==0.12.1
    # via
    #   black
    #   mypy
peewee==3.18.2
    # via yfinance
platformdirs==4.5.0
    # via
    #   black
    #   yfinance
pluggy==1.6.0
    # via pytest
protobuf==6.33.0
    # via yfinance
pycparser==2.23
    # via cffi
pydantic==2.12.3
    # via fastapi
pydantic-core==2.41.4
    # via pydantic
pygments==2.19.2
    # via pytest
pytest==8.4.2
    # via
    #   ai_agents (pyproject.toml)
    #   pytest-asyncio
pytest-asyncio==1.2.0
    # via ai_agents (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via uvicorn
pytokens==0.2.0
    # via black
pytz==2025.2
    # via
    #   pandas
    #   yfinance
pyyaml==6.0.3
    # via uvicorn
requests==2.32.5
    # via yfinance
ruff==0.5.0
    # via ai_agents (pyproject.toml)
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.8
    # via beautifulsoup4
starlette==0.48.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   pytest-asyncio
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.37.0
    # via ai_agents (pyproject.toml)
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via
    #   uvicorn
    #   yfinance
yfinance==0.2.66
    # via ai_agents (pyproject.toml)
